lib/libstd/Types: change ProcessID type to 32-bit unsigned integer

+++ lib/libstd/Types.h
@@ -137,7 +137,7 @@ typedef unsigned short UserID;
 typedef unsigned short GroupID;
 
 /** Process Identification Number. */
-typedef u16 ProcessID;
+typedef u32 ProcessID;
 
 /**
  * Describes a device ID number.
